services/bitfriendly/app/pyproject.toml
[tool.poetry]
name = "bitfriendly"
version = "0.1.0"
description = "bitcoin ord"
authors = ["Sech"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.10"
annotated-types = "0.6.0"
anyio = "3.7.1"
asn1crypto = "1.5.1"
astropy  = "5.3.4"
attrs = "23.2.0"
autoflake = "2.2.1"
base58 = "2.1.1"
base58check = "1.0.2"
base58grs = "2.1.1"
bcrypt = "4.1.2"
beautifulsoup4 = "4.12.3"
bech32 = "1.2.0"
bech32m = "1.0.0"
bech32m-chia = "1.0.3"
bip32 = "3.4"
bip32utils = "0.3.post4"
bit = "0.8.0"
bitcoin = "1.1.39"
bitcoin-utils = "0.6.5"
bitcoin-value = "1.6.0"
bitcoinX = "0.9"
blockcypher = "1.0.91"
btc-address-dump = "0.4.2"
btclib = "2023.7.12"
btclib_libsecp256k1 = "0.4.0"
cashaddress = "1.0.6"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
coincurve = "18.0.0"
cryptography = "41.0.7"
cryptotools = { git = "https://github.com/mcdallas/cryptotools.git", rev = "dbdced863bd2622f4cb41595df1ecad399760d78" }
docstring-parser = "0.15"
ecdsa = "0.17.0"
electrumsv-secp256k1 = "18.0.0"
exceptiongroup = "1.2.0"
fastapi = "0.104.1"
greenlet = "3.0.3"
groestlcoin_hash = "1.0.3"
gunicorn = "21.2.0"
h11 = "0.14.0"
h5py = "3.9.0"
hdwallet = "2.2.1"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.25.2"
idna = "3.6"
install = "1.3.5"
mnemonic = "0.21"
mpmath = "1.3.0"
packaging = "23.2"
passlib = "1.7.4"
pyasn1 = "0.5.1"
pycoin = "0.92.20230326"
pycparser = "2.21"
pycryptodome = "3.20.0"
pycryptodomex = "3.20.0"
pydantic = "2.6.0"
pydantic-settings = "2.1.0"
pydantic_core = "2.16.1"
pyerfa = "2.0.1.1"
pyflakes = "3.2.0"
PyJWT = "2.8.0"
pysqlite3 = "0.5.2"
python-bitcoinrpc = "1.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyuvdata = "2.4.1"
PyYAML = "6.0.1"
requests = "2.31.0"
rsa = "4.9"
scrypt = "0.8.21"
setuptools-scm = "8.0.4"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
SQLAlchemy = "2.0.25"
starlette = "0.27.0"
sympy = "1.12"
tomli = "2.0.1"
typing_extensions = "4.9.0"
urllib3 = "2.1.0"
uvicorn = "0.23.2"
uvtools = "0.2.0"
numpy = "1.22.0"


[[tool.poetry.source]]
name = "cryptotools"
url = "https://github.com/mcdallas/cryptotools.git"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
